Updated: Questions & CreateGroup (& JSONWebDB)
Questions load an explanation how to add new questions when questionlist
is empty.
CreateGroups is fully implemented now, user is added to the group
imedeatly after creating it. (Which is expected)
(According functions have been changed in the JSONWebDB)
HeaderActivity now implements the share/admin button, and header_black
includes it. Since it's only nessecairy in activities which implement
the header_black.
Added the share button and a few strings, check the file for specifics